+++ src/extensions/mvs/components/annotation-label/visual.ts
@@ -15,9 +15,9 @@ import { VisualContext } from '../../../../mol-repr/visual';
 import { Theme } from '../../../../mol-theme/theme';
 import { arrayEqual } from '../../../../mol-util';
 import { ColorNames } from '../../../../mol-util/color/names';
-import { FormatTemplate } from '../../../../mol-util/formatting';
 import { omitObjectKeys } from '../../../../mol-util/object';
 import { ParamDefinition as PD } from '../../../../mol-util/param-definition';
+import { FormatTemplate } from '../../../../mol-util/string-format';
 import { textPropsForSelection } from '../../helpers/label-text';
 import { MVSAnnotationRow } from '../../helpers/schemas';
 import { GroupedArray } from '../../helpers/utils';
